Grouping and filtering of inspection DataFrames
There is a function introduced for grouping and filtering of the inspection DataFrames. The function allows
for quite a generic usage, including hierarchical indexing, filtering of rows and columns and querying a DataFrame.

Example usage:

```python
filter_inspection_dataframe(
    df,
    groupby=["platform", "ncpus"],  # both `platform` and `ncpus` are shortened column names
    like="duration",  # a `regex` parameter can be used instead to filter by regex expression
    where="ncpus == 32 | ncpus == 64"  # note that the ncpus is a shortened column name
)
```

The convience of this function when filtering is that we don't need to provide full column names,
which are usually quite long in case of the inspection DataFrames.

Signed-off-by: Marek Cermak <macermak@redhat.com>

modified:   notebooks/development/Amun InspectionRun Analysis .ipynb
modified:   notebooks/development/Amun InspectionRun Analysis .py